+++ src/bspinfo.cpp
@@ -97,7 +97,7 @@ int parseArgs ( int index, const char *argv [] )
         if ( localError ) {
             errors = true;
             int offset = ptr - localCopy - 1;
-            size_t width = strlen ( ptr ) + 1;
+            int width = (int) strlen ( ptr ) + 1;
             fprintf ( stderr, "Unrecognized parameter '%*.*s'\n", width, width, argv [index] + offset );
         }
         free ( localCopy );
